Move terms to a separate module in the kernel
